Move history items details view models and views to Details folder
