+++ WalletWasabi.Fluent/ViewModels/Wallets/Home/History/Details/CoinJoinsDetailsViewModel.cs
@@ -10,22 +10,22 @@ using WalletWasabi.Fluent.Extensions;
 using WalletWasabi.Fluent.ViewModels.Navigation;
 using WalletWasabi.Fluent.ViewModels.Wallets.Home.History.HistoryItems;
 
-namespace WalletWasabi.Fluent.ViewModels.Wallets.Home.History;
+namespace WalletWasabi.Fluent.ViewModels.Wallets.Home.History.Details;
 
 [NavigationMetaData(Title = "Coinjoin Details")]
-public partial class CoinJoinDetailsViewModel : RoutableViewModel
+public partial class CoinJoinsDetailsViewModel : RoutableViewModel
 {
-	private readonly CoinJoinHistoryItemViewModel _coinJoin;
+	private readonly CoinJoinsHistoryItemViewModel _coinJoinGroup;
 
 	[AutoNotify] private string _date = "";
 	[AutoNotify] private string _status = "";
 	[AutoNotify] private Money? _coinJoinFee;
 	[AutoNotify] private string _coinJoinFeeString = "";
-	[AutoNotify] private uint256? _transactionId;
+	[AutoNotify] private ObservableCollection<uint256>? _transactionIds;
 
-	public CoinJoinDetailsViewModel(CoinJoinHistoryItemViewModel coinJoin)
+	public CoinJoinsDetailsViewModel(CoinJoinsHistoryItemViewModel coinJoinGroup)
 	{
-		_coinJoin = coinJoin;
+		_coinJoinGroup = coinJoinGroup;
 
 		SetupCancel(enableCancel: false, enableCancelOnEscape: true, enableCancelOnPressed: true);
 		NextCommand = CancelCommand;
@@ -48,7 +48,7 @@ public partial class CoinJoinDetailsViewModel : RoutableViewModel
 		base.OnNavigatedTo(isInHistory, disposables);
 
 		Observable
-			.FromEventPattern(_coinJoin, nameof(PropertyChanged))
+			.FromEventPattern(_coinJoinGroup, nameof(PropertyChanged))
 			.Throttle(TimeSpan.FromMilliseconds(100))
 			.Subscribe(_ => Update())
 			.DisposeWith(disposables);
@@ -56,11 +56,11 @@ public partial class CoinJoinDetailsViewModel : RoutableViewModel
 
 	private void Update()
 	{
-		Date = _coinJoin.DateString;
-		Status = _coinJoin.IsConfirmed ? "Confirmed" : "Pending";
-		CoinJoinFee = _coinJoin.OutgoingAmount;
+		Date = _coinJoinGroup.DateString;
+		Status = _coinJoinGroup.IsConfirmed ? "Confirmed" : "Pending";
+		CoinJoinFee = _coinJoinGroup.OutgoingAmount;
 		CoinJoinFeeString = CoinJoinFee.ToFeeDisplayUnitString() ?? "Unknown";
 
-		TransactionId = _coinJoin.CoinJoinTransaction.TransactionId;
+		TransactionIds = new ObservableCollection<uint256>(_coinJoinGroup.CoinJoinTransactions.Select(x => x.TransactionId));
 	}
 }
