+++ WalletWasabi.Fluent/Views/Wallets/Home/History/Details/CoinJoinDetailsView.axaml.cs
@@ -1,11 +1,11 @@
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace WalletWasabi.Fluent.Views.Wallets.Home.History;
+namespace WalletWasabi.Fluent.Views.Wallets.Home.History.Details;
 
-public class TransactionDetailsView : UserControl
+public class CoinJoinDetailsView : UserControl
 {
-	public TransactionDetailsView()
+	public CoinJoinDetailsView()
 	{
 		InitializeComponent();
 	}
