+++ WalletWasabi.Fluent/Views/Wallets/Home/History/Details/CoinJoinsDetailsView.axaml.cs
@@ -1,7 +1,7 @@
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace WalletWasabi.Fluent.Views.Wallets.Home.History;
+namespace WalletWasabi.Fluent.Views.Wallets.Home.History.Details;
 
 public class CoinJoinsDetailsView : UserControl
 {
